fix warning

+++ Common/GmshSocket.h
@@ -363,7 +363,7 @@ class GmshServer : public GmshSocket{
 #if !defined(WIN32) || defined(__CYGWIN__)
       if(tmpsock < 0)
 #else
-      if(tmpsock == INVALID_SOCKET)
+	if(tmpsock == (int)INVALID_SOCKET)
 #endif
         throw "Couldn't create socket";
       // bind the socket to its name
